remove temporary code, fix for loglevel
